Fixed bug with get patch
The issue I encountered was with rotate spitting out zeros if it
was fed a uint8 numpy array. A conversion to float sorted that.

+++ agent.py
@@ -42,7 +42,8 @@ class Agent(threading.Thread):
         # rotate's angle argument is how many degrees to rotate *anticlockwise*
         # Additionally, the coordinates are reversed in skimage
         ### pixels is the game board oriented in the direction of the agent ###
-        pixels = pad(state.pixels, hps+1, mode='constant', constant_values=1)
+        pixels = pad(state.pixels.astype('float'), hps+1, 
+                     mode='constant', constant_values=1)
         # Because of padding x0 <- x0 + hps + 1
         x0, x1 = x0+hps+1, x1+hps+1
         pixels_rot = rotate(image=pixels, angle=angle,
